update: move sine realated and add factory
