+++ src/py/mat3ra/made/tools/utils/factories.py
@@ -1,0 +1,7 @@
+from mat3ra.utils.factory import BaseFactory
+
+
+class PerturbationFunctionHelperFactory(BaseFactory):
+    __class_registry__ = {
+        "sine_wave": "mat3ra.made.tools.utils.helpers.SineWave",
+    }
